restart: add all files from prev. project
The previous quiz app project (by YT Daily
Tuition) had issues with running as it was
generated with an outdated create-react-app.
It is recommended as of 2023 that create-react-app
is not used for projects. So, instead I built
this app manually. I added and configured
Webpack and Babel by myself, including react,
react-dom, and react config. using the latest
version.